chore: made some version-related functions shared
